requirements.txt
python-ev3dev2
pybricks
numpy